Refactor validator package into modular files
