+++ validator/validators_map.go
@@ -1,0 +1,89 @@
+package validator
+
+import "sync"
+
+var validators = map[string]ValidationFunc{
+	"required":      requiredValidator,
+	"max":           maxValidator,
+	"min":           minValidator,
+	"range":         rangeValidator,
+	"email":         emailValidator,
+	"regex":         regexValidator,
+	"numeric":       numericValidator,
+	"alpha":         alphaValidator,
+	"alphanum":      alphanumValidator,
+	"alphaspace":    alphaSpaceValidator,
+	"aspace":        alphaSpaceValidator,
+	"alphaspacenum": alphaSpaceNumValidator,
+	"aspacenum":     alphaSpaceNumValidator,
+	"startswith":    startsWithValidator,
+	"starts":        startsWithValidator,
+	"endswith":      endsWithValidator,
+	"ends":          endsWithValidator,
+	"contains":      containsValidator,
+	"notcontains":   notContainsValidator,
+	"notcont":       notContainsValidator,
+	"ascii":         asciiValidator,
+	"positive":      positiveValidator,
+	"pos":           positiveValidator,
+	"negative":      negativeValidator,
+	"neg":           negativeValidator,
+	"even":          evenValidator,
+	"odd":           oddValidator,
+	"multiple":      multipleValidator,
+	"mult":          multipleValidator,
+	"base64":        base64Validator,
+	"b64":           base64Validator,
+	"json":          jsonValidator,
+	"semver":        semverValidator,
+	"extension":     extensionValidator,
+	"ext":           extensionValidator,
+	"url":           urlValidator,
+	"ip":            ipValidator,
+	"ipv4":          ipv4Validator,
+	"ipv6":          ipv6Validator,
+	"domain":        domainValidator,
+	"mac":           macValidator,
+	"port":          portValidator,
+	"date":          dateValidator,
+	"pastdate":      pastDateValidator,
+	"past":          pastDateValidator,
+	"futuredate":    futureDateValidator,
+	"future":        futureDateValidator,
+	"workday":       workdayValidator,
+	"wday":          workdayValidator,
+	"weekend":       weekendValidator,
+	"wend":          weekendValidator,
+	"in":            inValidator,
+	"notin":         notInValidator,
+	"length":        lengthValidator,
+	"len":           lengthValidator,
+	"between":       betweenValidator,
+	"btw":           betweenValidator,
+	"boolean":       booleanValidator,
+	"bool":          booleanValidator,
+	"uuid":          uuidValidator,
+	"creditcard":    creditCardValidator,
+	"cc":            creditCardValidator,
+	"eq":            equalValidator,
+	"ne":            notEqualValidator,
+	"lt":            lessThanValidator,
+	"lte":           lessThanOrEqualValidator,
+	"gt":            greaterThanValidator,
+	"gte":           greaterThanOrEqualValidator,
+	"realemail":     realEmailValidator,
+	"remail":        realEmailValidator,
+	"password":      passwordValidator,
+	"pwd":           passwordValidator,
+	"phone":         phoneValidator,
+	"username":      usernameValidator,
+	"uname":         usernameValidator,
+	"slug":          slugValidator,
+	"hexcolor":      hexcolorValidator,
+	"hcolor":        hexcolorValidator,
+	"fullname":      fullnameValidator,
+	"fname":         fullnameValidator,
+	"name":          nameValidator,
+}
+
+var validatorsMutex sync.RWMutex
